refactor: add tests as pytests
