--- requirements.txt
@@ -1,1 +1,0 @@
-msc-pyparser>=1.2.1
