+++ tests/test_linter.py
@@ -1,0 +1,5 @@
+from crs_linter.linter import Check
+
+def test_check(data, txvars):
+    c = Check(data, txvars)
+    assert c.check_ignore_case() == True
